# swagger.yaml
openapi: 3.0.0
info:
  title: WhSmith
  contact: {}
  version: '1.0'
servers:
- url: https://avnology.org
  variables: {}
- url: http://localhost:1812
  variables: {}
paths:
...
  /role/add:
    post:
      tags:
      - Roles
      summary: Add Role
      operationId: AddRole
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AddRoleRequest'
              - example:
                  name: 'Manger '
                  permissions:
                  - manage order
                  - manage category
            example:
              name: 'Manger '
              permissions:
              - manage order
              - manage category
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /role/remove/{id}:
    delete:
      tags:
      - Roles
      summary: remove role
      operationId: removerole
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /role/update/{id}:
    patch:
      tags:
      - Roles
      summary: Update Role
      operationId: UpdateRole
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateRoleRequest'
              - example:
                  role: manage user
            example:
              role: manage user
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /role/all:
    get:
      tags:
      - Roles
      summary: Get All Roles
      operationId: GetAllRoles
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
components:
  schemas:
...
    AddRoleRequest:
      title: AddRoleRequest
      required:
      - name
      - permissions
      type: object
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
          description: ''
      example:
        name: 'Manger '
        permissions:
        - manage order
        - manage category
    UpdateRoleRequest:
      title: UpdateRoleRequest
      required:
      - role
      type: object
      properties:
        role:
          type: string
      example:
        role: manage user
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...